Fix imports and centralize exceptions
