+++ StrategyFactory.py
@@ -8,6 +8,7 @@ from data_feed import DataFeed
 from indicators import TechnicalIndicators
 from config import get_config
 from utils import handle_error
+from exceptions import DataError, StrategyError
 
 logger = logging.getLogger(__name__)
