+++ service/market_data_service.py
@@ -4,11 +4,9 @@ from service.repository_service import RepositoryService, transaction
 from database.market_data_repository import MarketDataRepository
 from utils import handle_error
 import logging
+from exceptions import BaseTradingException, DataError, MarketDataServiceException
 
 logger = logging.getLogger(__name__)
-
-class MarketDataServiceException(BaseTradingException):
-    pass
 
 class MarketDataService(abc.ABC):
     @abc.abstractmethod
